🦈 IMP: LMR - Reduce Less if Checking
--- STC ---
ELO   | 2.40 +- 1.52 (95%)
SPRT  | 10.0+0.10s Threads=1 Hash=16MB
LLR   | 2.99 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 103872 W: 27457 L: 26740 D: 49675

--- LTC ---
ELO   | 2.65 +- 1.98 (95%)
SPRT  | 60.0+0.60s Threads=1 Hash=256MB
LLR   | 2.98 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 56448 W: 13684 L: 13253 D: 29511

Bench: 9959615

+++ src/Engine/LogarithmicReductionTable.h
@@ -18,10 +18,10 @@ namespace StockDory
     {
 
         private:
-            static std::array<std::array<uint8_t, MaxMove>, MaxDepth> Internal;
+            static std::array<std::array<int16_t, MaxMove>, MaxDepth> Internal;
 
         public:
-            static inline uint8_t Get(const uint8_t depth, const uint8_t move)
+            static inline int16_t Get(const uint8_t depth, const uint8_t move)
             {
                 return Internal[depth][move];
             }
@@ -30,11 +30,11 @@ namespace StockDory
 
 } // StockDory
 
-std::array<std::array<uint8_t, MaxMove>, MaxDepth> StockDory::LogarithmicReductionTable::Internal = []() {
-    std::array<std::array<uint8_t, MaxMove>, MaxDepth> temp = {};
+std::array<std::array<int16_t, MaxMove>, MaxDepth> StockDory::LogarithmicReductionTable::Internal = []() {
+    std::array<std::array<int16_t, MaxMove>, MaxDepth> temp = {};
 
     for (uint8_t depth = 1; depth < MaxDepth; depth++) for (uint8_t move = 1; move < MaxMove; move++) {
-        temp[depth][move] = static_cast<uint8_t>(log(depth) * log(move) / 2 - 0.2);
+        temp[depth][move] = static_cast<int16_t>(log(depth) * log(move) / 2 - 0.2);
     }
 
     return temp;
